Tighten up the conditional guard in get/set methods
Makes it so that if a value that is not Undefined
gets passed in, it gets assigned to the ivar. If
no value (or Undefined) gets passed in, the ivar
will be returned.

Because of the use of DataMapper::Mapper.model in
DataMapper::Mapper.inherited, I had to move the
toplevel require statements below the toplevel
DataMapper module definition.