Don't run mutant on travis for now
The lib is *far* from mutation coverage and running
mutant on travis atm is an utter waste of travis'
processing power and greatly increases the time we
have to wait for ci results.

The few people working on dm-mapper can and should
run mutant locally, but imho running it on travis
right now, just isn't worth it.